Add files via upload
Spreadsheet Analysis project to analyse the data in the sales csv dataset. The  output of data analysis is depicted in the pie chart and bar plot. The analysis on monthly percentage change, max and min sale months, total sales as list, average on sales/year are displayed in the csv files called results.csv.